jobs/batch-permit-validator/pyproject.toml
[tool.poetry]
name = "batch-permit-validator"
version = "1.0.1"
description = ""
authors = []
packages = [{include = "batch_permit_validator", from = "src"}]
license = "BSD 3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
flask = "^3.0.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
launchdarkly-server-sdk = "^9.0.1"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.44.1"
flask-babel = "^4.0.0"
gunicorn = "^21.2.0"
pg8000 = "^1.31.2"
gcp-queue = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/gcp-queue", branch = "main" }
structured-logging = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/structured-logging", branch = "main" }
# strr-api = {git = "https://github.com/bcgov/STRR.git", branch = "feature-validator-updates", subdirectory = "strr-api"}
strr-api = {git = "https://github.com/bcgov/STRR.git", rev = "main", subdirectory = "strr-api"}
nanoid = "^2.0.0"
redis = "^5.2.1"

[tool.poetry.group.test.dependencies]
freezegun = "^1.4.0"
hypothesis = "^6.100.0"
pexpect = "^4.9.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
faker = "^24.4.0"
isort = "^5.13.2"
black = "^24.10.0"
pylint = "^3.0.3"
testcontainers = "^4.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "google"
line_length = 120
known_first_party = ["strr_email"]


[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.pylint]
max-line-length = 132

[tool.pytest.ini_options]
minversion = "2.0"
testpaths = [
   "tests",
]
addopts = "--verbose --strict -p no:warnings --cov=src --cov-report html:htmlcov --cov-report xml:coverage.xml"
python_files = [
   "test*.py"
]
norecursedirs = [
   ".git", ".tox", "venv*", "requirements*", "build",
]
log_cli = true
log_cli_level = "1"
filterwarnings = [
   "ignore::UserWarning"
]
markers = [
   "slow",
   "serial",
   "conf: Inject flask config values into the app fixture",
   "load: mark test as a heavy load test to exclude from standard CI runs",
]
